+++ tests/S3Action_test.cpp
@@ -19,8 +19,8 @@ SCENARIO("Calculate the bulk action on S3 triangulations", "[action]")
 {
   GIVEN("A 3D 2-sphere foliated triangulation.")
   {
-    constexpr auto simplices  = static_cast<int_fast32_t>(6400);
-    constexpr auto timeslices = static_cast<int_fast32_t>(7);
+    constexpr auto simplices  = static_cast<int_fast64_t>(6400);
+    constexpr auto timeslices = static_cast<int_fast64_t>(7);
     constexpr auto K          = static_cast<long double>(1.1);
     constexpr auto Lambda     = static_cast<long double>(0.1);
     Manifold3      universe(simplices, timeslices);
